test/fixtures/wpt/dom/events/scrolling/scrollend-event-fired-after-sequence-of-scrolls.tentative.html

<!DOCTYPE HTML>
<meta name="viewport" content="width=device-width,initial-scale=1,minimum-scale=1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="scroll_support.js"></script>
<style>
#targetDiv {
  width: 200px;
  height: 200px;
  overflow: scroll;
}

#innerDiv {
  width: 500px;
  height: 4000px;
}
</style>

<body style="margin:0" onload=runTest()>
<div id="targetDiv">
  <div id="innerDiv">
  </div>
</div>
</body>

<script>
const target_div = document.getElementById('targetDiv');
let scrollend_arrived = false;
let scrollend_event_count = 0;

function onScrollEnd(event) {
  assert_false(event.cancelable);
  assert_false(event.bubbles);
  scrollend_arrived = true;
  scrollend_event_count += 1;
}

function runTest() {
  promise_test (async (t) => {
    // Make sure that no scrollend event is sent to document.
    document.addEventListener("scrollend",
        t.unreached_func("document got unexpected scrollend event."));
    await waitForCompositorCommit();

    // Scroll down & up & down on target div and wait for the target_div to get scrollend event.
    target_div.addEventListener("scrollend", onScrollEnd);
    const move_path = [
      { x: 0, y: -300}, // down
      { x: 0, y: -100}, // up
      { x: 0, y: -400}, // down
      { x: 0, y: -200}, // up
    ];
    await touchScrollInTargetSequentiallyWithPause(target_div, move_path, 150);

    await waitFor(() => {return scrollend_arrived;},
        'target_div did not receive scrollend event after sequence of scrolls on target.');
    assert_equals(scrollend_event_count, 1);
  }, "Move down, up and down again, receive scrollend event only once");
}
</script>
